flow_pre_clean
Cleaned out unused and dead code.
