+++ src/main/scala/Dispatcher.scala
@@ -12,7 +12,7 @@ object Dispatcher {
         case "dns_post_lda" => dns_post_lda.run()
         case "dns_pre_lda" => dns_pre_lda.run()
         case "flow_post_lda" =>flow_post_lda.run()
-        case "flow_pre_lda" => flow_pre_lda.run()
+        case "flow_pre_lda" => FlowPreLDA.run()
       }
 
       System.exit(0)
